requirements.txt
ipython==8.12.3
Pillow==10.2.0
protobuf==5.26.0
python-dotenv==1.0.1
streamlit==1.32.2
